refactor(p1): extract CalendarAccessTools.kt — calendar_list / calendar_create + timeMillis
Closes Group A (Contacts/Sms/Phone/Calendar). Same factory shape; in addition,
JsonElement.timeMillis (only consumed by createCalendarEvent) leaves the
coordinator and lives file-private next to its sole caller.

Side cleanup in SystemAccessTools.kt:
- drop dead imports `android.provider.CalendarContract`, `java.time.Instant`, `java.time.ZoneId`

SystemAccessTools.kt: 1441 → 823 lines (-43% so far across Group A).

Verification: :app:compileDebugKotlin --offline → BUILD SUCCESSFUL.
